Extract statsd URL from main analytics file
We want to be able to use statsd on the server-side. However,
lib/analytics cannot be used on the server-side because of its
dependence on the DOM. Extracting statsd to a module that does not
depend on the DOM will allow us to use it on the server.
